+++ _tests/tests/unit_tests/iotd.php
@@ -24,6 +24,12 @@ class iotd_test_set extends ocp_test_case
 	{
 		parent::setUp();
 
+		if (in_safe_mode())
+		{
+			$this->assertTrue(false,'Cannot work in safe mode');
+			return;
+		}
+
 		require_code('iotds');
 
 		$this->iotd_id=add_iotd('http://google.com','welcome','Google','images/google.jpg',0,0,0,0,'Notes ?',NULL,NULL,0,NULL,0,NULL);
@@ -33,6 +39,12 @@ class iotd_test_set extends ocp_test_case
 
 	function testEditIotd()
 	{
+		if (in_safe_mode())
+		{
+			$this->assertTrue(false,'Cannot work in safe mode');
+			return;
+		}
+
 		edit_iotd($this->iotd_id,'Thank you','Caption ?','images/yahoo.jpg','yahoo.com',0,0,0,'Notes');
 
 		$this->assertTrue('yahoo.com'==$GLOBALS['SITE_DB']->query_value('iotd','url',array('id'=>$this->iotd_id)));
@@ -40,6 +52,12 @@ class iotd_test_set extends ocp_test_case
 
 	function tearDown()
 	{
+		if (in_safe_mode())
+		{
+			$this->assertTrue(false,'Cannot work in safe mode');
+			return;
+		}
+
 		delete_iotd($this->iotd_id);
 		parent::tearDown();
 	}
